Move PingPongClient to its own file
